Fix : noti switch
스위치 위치 바꿈 + 백엔드 연결 준비
